refactor: replace remaining any types with proper TypeScript types
- Replace any Serie type with SerieData from @/lib/series/types in PostLayout
- Create shared Serie and SeriePost interfaces in series types
- Update SeriePlaylist to use shared Serie interface
- Replace any AuthorsMap types with proper AuthorsMap import
- Update experiments page to use ExperimentProps[] instead of any[]
- Replace Video any types with proper Video interface
- Update videos page and VideoCard to use shared Video types
- Update ContentListing renderItem to use ContentItem type
- Fix type compatibility issues with exactOptionalPropertyTypes
- Handle nullable slugs and optional video/slides properties
- Improve type safety across components and pages
