+++ components/VideoCard/types.ts
@@ -1,8 +1,21 @@
-export interface VideoCardProps {
-  title: string
-  description?: string
+interface VideoThumbnail {
   url: string
-  thumbnail?: string
-  duration?: string
-  publishedAt?: string
+}
+
+export interface Video {
+  id: string
+  title: string
+  description: string
+  thumbnails: {
+    default: VideoThumbnail
+    medium?: VideoThumbnail
+    high?: VideoThumbnail
+    standard?: VideoThumbnail
+    maxres?: VideoThumbnail
+  }
+}
+
+export interface VideoCardProps {
+  video: Video
+  playButton?: boolean
 }
