+++ lib/series/types.ts
@@ -1,5 +1,16 @@
 import type { FrontMatter } from '../../types'
 
+export interface SeriePost {
+  slug: string
+  title: string
+}
+
+export interface Serie {
+  title: string
+  posts: SeriePost[]
+  slug: string
+}
+
 export interface SerieData {
   title: string
   description?: string
